pyproject.toml
[tool.poetry]
name = "rndmesnd"
version = "0.1.0"
description = ""
authors = ["istimaldar <istimaldar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
dynaconf = "^3.1.4"
python-telegram-bot = "^13.3"
pony = "^0.7.14"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
